service do exchange com mais responsabilidades, deixando o controller mais solto, busca de historico do usuario criado.
