+++ api/.env.example
@@ -59,3 +59,5 @@ VITE_PUSHER_APP_CLUSTER="${PUSHER_APP_CLUSTER}"
 
 WWWGROUP=1000
 WWWUSER=1000
+
+SAIL_XDEBUG_MODE=develop,debug,coverage
